requirements.txt
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
drf-yasg==1.20.0
openpyxl==3.0.9
pandas==1.3.4
